refactor(backend): inline evlog Better Auth middleware in app.ts
Remove betterAuthIdentify.ts; keep createAuthMiddleware next to createApp().
